+++ src/models/order_item.rs
@@ -1,4 +1,4 @@
-
+use rust_decimal::Decimal;
 #[derive(serde::Serialize, serde::Deserialize, Debug, Clone, PartialEq, sqlx::FromRow)]
 
 pub struct OrderItem {
@@ -6,5 +6,5 @@ pub struct OrderItem {
     pub order_id: Option<i32>,
     pub product_id: Option<i32>,
     pub quantity: Option<i16>,
-    pub price: Option<f32>,
+    pub price: Option<Decimal>,
 }
